pyproject.toml
[tool.poetry]
name = "django-blog"
version = "0.1.0"
description = ""
authors = ["TurkaljGoran <goran@iconis.digital>"]
readme = "README.md"
packages = [{include = "django_blog"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
django-crispy-forms = "^2.1"
crispy-bootstrap4 = "^2023.1"
pillow = "^10.1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
